scripts/requirements.txt
# AI Daily - Python Dependencies
tavily==0.4.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0